common/flash: Aligns persistent data to Flash Size
Some platforms have allocated a region in the flash to store
persistent data. Platforms with error correction code can require
writes to occur in blocks. CONFIG_FLASH_WRITE_SIZE defines a
valid block size for the platform.

BUG=b:149870870
TEST=Verified that loading and saving persistent data functions
on servoV4. Verified that the structure sizes are proportional
to N * sizeof(alignment)

Change-Id: I54a5506f2179c98cc7383266598b9074c3521b8c