usb_mux: retimer: mux as chained mux and retimer
This makes retimers appear as generic muxes.  By allowing a
chain of muxes they can be stacked up to the new configurations
that zork requires and will continue to work as they did before
on configurations that only have a single mux.

The code used to have two different arrays, 1) muxes and 2)
retimers.  On one of the zork configurations the processor
MUX stopped being the primary mux and the retimer took its
place.  In a different configuration of that same platform
it left the primary and secondary alone but the mux_set
FLIP operation had to be ignored.  Since the same
interfaces needed to be available for both it stopped making
sense to have two different structures and two different
methods of handling them.  This consolodates the two into
one.

The platforms that do not have retimers, this change will
not make any difference. For platforms like zork, it will
remove the retimers and make them chained muxes.  So
testing on trembyle makes sense to verify,

BUG=b:147593660
BRANCH=none
TEST=verify USB still works one trembyle

Change-Id: I286cf1e302f9bd3dd7e81098ec08514a2a009fe3
Signed-off-by: Denis Brockus <dbrockus@chromium.org>