dedede: Configure EN_VCCST as input
Currently, the EC isn't using EN_VCCST itself, but we are using
external logic to set this pin.  The pin is initialized to low on the
EC which causes a drive fight.

This commit changes GPIO_EN_VCCST to an input while we determine
whether or not we will use this pin in the EC.

BUG=b:149775160
BRANCH=None
TEST=`make -j buildall`
TEST=Build and flash waddledoo, verify EC cannot set EN_VCCST.

Change-Id: Ia9f439d3d049580ac66689fdde11ed75ca044584
Signed-off-by: Aseda Aboagye <aaboagye@google.com>