make: add preprocessor stage
For the upcoming introduction of transitioning Cr50 console
communications to packet mode, there is a need to be able to replace
all print function invocations in the code with calls to packet
sending function.

This replacement is easiest to make in C preprocessor outputs, as
there all macros are replaced with actual function invocations.

This patch adds a configuration option CONFIG_EXTRACT_PRINTF_STRINGS,
when enabled, building of the image object files starts happening in
three steps instead of one, instead of .c => .o transition, the steps
are .c => .E => .Ep => .o, where .E is the C preprocessor output, and
.Ep is result of processing by ./util/util_precompile.py.

BUG=b:149964350
TEST=image layout does no change if CONFIG_EXTRACT_PRINTF_STRINGS is
     not defined. With the rest of the patches applied defining the
     above config option allows to build a Cr50 image supporting
     packet console communications mode.

Change-Id: I20b8ba7c5d13cb54ac6adbdbce856d92023ce997
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>