chip/nxcp/lpc: Fix race condition when reading 8042 data.
When reading NPCX_HIKMDI it deasserts the IBF status flag. This means
the AP is allowed to write to the input buffer. Because NPCX_HIKMDI is
read twice, that means the AP could have written to the input buffer
between reads. This results in losing one of the interrupts. This only
happens if DEBUG_LPC is enabled.

BUG=b:145575366
BRANCH=none
TEST=Verified we don't lose any writes by examining logs.

Change-Id: I2904c316fcad55001e8d297f4a0a73073b07702b
Signed-off-by: Raul E Rangel <rrangel@chromium.org>