script for processing preprocessor output
This script is customized to be used in Chrome OS EC codebase, where
functions cprintf(), cprints() and cputs()are used to generate console
output.

The operation of the script is extensively described in the docstring
of the file, a unit test is provided to verify basic functionality.

BRANCH=cr50, cr50-mp
BUG=b:149964350
TEST=./util/test_util_precompile.py succeeds with the rest of the
     patches applied built a Cr50 image, verified that using a
     verified version of Acroterm one can communicate with Cr50
     console the same way it works today over regular UART.

Change-Id: I2c92088a8b08dcf7fdb606a01fbc94ab0d2df6e6
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>