Update patch set 1
Patch Set 1: Code-Review+2

Thank you for catching this. In other boards we are not seeing it because the port is auto-initialized to 0 by depthcharge(because of partial struct initialization) and I2C port 0 has been used.

In the case here, as per the bug, port 0 is not used and hence the get_i2c_port fails. This CL will address the situation where port 0 is not used.

Patch-set: 1
Reviewer: Gerrit User 1302633 <1302633@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2