Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Wow this touches a lot of files! I wonder if there is some way to default to the old behavior to avoid changing the files that aren't using any chaining?

Most of the changes are to use a small amount of new code in usb_mux.c.   I tried to make the general handling of the usb_mux.c code do the original code, for the most part, when the chaining was not being used.  This made the code really hard to read.

Patch-set: 11
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0