Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Should we move all files from driver/retimer/* to driver/usb_mux/* ?
> I think that would make sense (but perhaps as follow-on CL).

Yes but it is hard to compare a file if it is just deleted in one location and then created in another.  This was the safest mechanism so apples and apples could be compared.

Patch-set: 11
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0