TCPMv2: Probe cable identity in PE_SRC_Discovery
This change adds probing of the cable identity during startup as a
source.  There is no timing requirement for message spacing when sent
from discovery, but this change starts with DiscoverIdentityTimer
spacing to avoid spamming the cable too quickly.  This can be tuned
later if desired.

When the port partner is running PD 2.0, the communications with the
cable must also use PD 2.0.  Otherwise, communications may be whatever
revision the cable responds with.

This change also corrects the tVDMBusy timer, which is 50 ms in both the
PD 2.0 and PD 3.0 specs.

BUG=b:148834626
BRANCH=None
TEST=on kindred, confirmed:
1. cable was probed during startup on connection as source
	a. cable probing stopped after an ACK
	b. cable probing did not continue past 20 messages with no
	GoodCRC
2. cable was note probed during startup as sink (charger connected)

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I65979f30d26ad8a37f507994a24ddc86c5cb41ff