Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > May I ask the reasoning for enabling it on these boards specifically?
> > 
> > I plan to add a Cr50 autotest, which runs uart_stres_tester.py.
> > It will check if there is any UART character drop. It will check if CR50 UART-USB bridge quality, but I think it can check CCD connection quality as well.
> > In the past, we enabled this flag for Hatch, puff, and kukui for this reason.
> 
> +@keithshort, @ecgh in CC, reviewers of similar patch for volteer and zork.
> 
> I am trying to enable this for all new baseboards, not just for dedede.

If you're trying to enable it to for all new baseboards, I wonder if you can enable it by default in include/config.h and just guard it with some conditional to match "a chrome OS device with Cr50". Off the top of my head, I was thinking about CONFIG_CHIPSET_* && CONFIG_USB_PD.

I'm fine with it either way, but was just trying to think of a way to reduce the number of CLs needed and to prevent us from forgetting in the future. Up to you.

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2