ec: enable CONFIG_CMD_CHARGEN as default
chargen is a cli command that generates a series of alphnumeric
characters. This command is useful in testing UART related modules
with a stress workload, for example CR50 UART-USB bridging.
This patch enables this command as default, so that boards can be
tested with UART-stress tester.

This patch excludes boards that do not need it, or that is not built
with this config.

This patch occupies 400 bytes or so in flash usage.

BUG=b:38448364
BRANCH=None
TEST=buildall

Change-Id: I6d1f3ec796ba9fe0a85c7372578eb10726e92be0
Signed-off-by: Namyoon Woo <namyoon@chromium.org>