Update patch set 1
Patch Set 1: Code-Review+2

I am OK with this or having the "probe" type checking removed.  This is the only driver that I have seen in the EC base that performs this type of verification.  So not sure if they should all have it, none should have it or if it is just this one that we need it in.

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2