compare_build.sh: Fix restore detached HEAD
This fixes the scenario where you want to run
compare_build.sh, but you are currently are in
a detached HEAD state.
Without it, ./compare_build.sh moves your HEAD
and cannot restore it.

The command "git rev-parse --abbrev-ref HEAD"
only returns a unique identifier (branch name)
when HEAD is on a branch. If you are in detached
HEAD mode, it simply returns HEAD, which isn't useful.

BRANCH=none
BUG=none
TEST=git checkout cros/master
     git rev-parse HEAD
     ./util/compare_build.sh
     git rev-parse HEAD
     # We should still be on the same hash

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: Ib8dbedc61a77567262fceab59c17e1fb9c78cdd6