Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> please, let's know is there any further comments on this, or we should go with patchset:7 which was last working with complete functionality.

I tried PS9 on Helios and it seems working as expected. EC took over the fan at 65 and kept controlling it until 80. Then, it seems DPTF reclaimed it and decided to shutdown the system. If DPTF were dead, EC would have successfully done the same.

To me, the rest looks a matter of selecting right thresholds for DPTF and EC in order to get the desired thermal control.

Which Hatch are you using? Why is the activation threshold so low? Can it be 60? Can you share the DPTF configuration of your device?

Patch-set: 9
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>