kodama: rewrite charger CHG_CTRL1 register
USB2.0 device cannot be recognized probabilistically when the smart
battery is activated with the adapter at the first time. It found
charger register CHG_CTRL1(0x11) is changed to 0x02 when disconnecting
the external device since the unstable Vbat. In fact, the normal value
should be 0x10.
Rewriting the charger CHG_CTRL1 register after jumping to RW.

BUG=b:150515707
TEST=ensure USB2.0 device can be recognized after the smart battery
     is activated with the adapter at the first time.
BRANCH=kukui

Signed-off-by: Xiong Huang <xiong.huang@bitland.corp-partner.google.com>
Change-Id: Iccba1f81099575968faa26bd8686c5ede37a3f31