max14637: Modify driver to have chip_enable active by default
The max14637 provides client mode bc1.2 detection, and the driver
enables bc1.2 detection or turns off the IC bases on VBUS
changes. However, the driver does not current differentiate between
VBUS resulting for the port acting as a sink or source. The result is
that bc1.2 detection is also tirggered when the port is acting as a
source. This can cause interop issues with other devices which aren't
expecting bc1.2 detection singaling in this case.

This CL modifies the driver so that the max14637 enable is on by
default. This results in the bc1.2 driver being mostly a NOP when the
port power role is a source. If the port power role is a sink, then
the enable line is pulsed low so that bc1.2 client detection is
triggered as expected.

BUG=b:147833952
BRANCH=firmware-hatch-12672.B
TEST=Verfied that the device which previously was not enumerating,
does enumerate. Test both legacy chargers to ensure that client bc1.2
works as expected.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I8bfd243e312c7238a6668edcdfcde2ac0ae5d580
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2086731
Reviewed-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
(cherry picked from commit e8dc03b3cb24df99f41311e842a5641425cfa871)