dratini: Enable/disable HDMI and debounce HPD
Dratini had the EN_HDMI signal previously routed to PCH_GATE_C10_L,
which is still asserted in runtime S0ix, which would cause the HPD to
pulse off temporarily, and then back on. This meant an HDMI monitor
would briefly power off in those situations, then back on, because it
looked like an HDMI monitor was plugged in again. Based on jwp@'s rework
(and thus the next board revision), which hands control of EN_HDMI to
the EC, this CL enables and disables HDMI based on the power state of
the AP. It also debounces the HPD signal before enabling/disabling
MST hub power.

BUG=b:147261818
BRANCH=firmware-hatch-12672.B
TEST=Verified by ODM

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: Ifb30161222fb7f27d13211b6c4bae616113cc7fb