bq25710: Do not soft reset when running out of RW image
The bq25710 does not have a reset pin it's only following a power on
reset that its registers are reset. The driver's init function
executes a soft reset of the bq25710 to ensure that register settings
following either an EC reboot or power on reset are the same.

However, this is happening when the EC is in RO or RW, and there is
not need to do this in RW. This CL adds a check around the soft reset
so that it's only done in R0.

BUG=b:148189096
BRANCH=firmware-hatch-12672.B
TEST=Verified that low power mode no longer gets reenabled following
the jump to RW.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I78a7484d3e44147cf350b785889f0f913b03bd06