driver: bma2x2: Fix frequency precision error
On nautilus, the max frequency is set to 62.5Hz (the EC max frequency is
set to 100Hz).
However, when setting frequency to 62500 (mHz), the EC would bump up the
frequency to 125Hz.
The problem was in the calculation of the true frequency.
By using 7.81Hz instead of 7.8125Hz as the multiplier, the driver would
calculate the true frequency as 62480mHz, so it would set the
accelerometer frequency to the next data rate.

BUG=chromium:1059456
BRANCH=poppy
TEST=On nocturne, check frequency is set right and
tast run <IP> hardware.SensorRing passes.

Change-Id: I743201bc8cccc7d38f39d1e370d660f4f37b4d95
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>