helios: Add custom battery presence detect
The BQ40Z50 seems to have a quirk around coming back from its
disconnected state. If the battery is physically present, and is coming
back from cut-off, we check for the function of the charge FET and the
disconnect FET first, before declaring that the battery is actually
present.  This prevents brownouts from happening when the first AP
G3->S0 sequence happens after cutoff, because the battery was not yet
actually ready to supply energy to the system.

BUG=b:141577278
BRANCH=firmware-hatch-12672.B
TEST=Verify DUT can boot from cutoff to S0 successfully on the first
try. Verified by ODM by same test on several SKUs.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I33ec4ea61a70ec8a115972e621ec12c4f97767d7