dedede: Require >15W to boot without a battery
Waddledoo appears to have issues booting with just a 15W supply.
Therefore, this commit sets the minimum power required to boot to be
greater than 15W such that a PD contract is required. Otherwise, if the
battery is present and has greater than 1% SOC, then booting is
allowed.

BUG=b:150702984
BRANCH=None
TEST=build and flash waddledoo, plug in a 15W non PD charger, verify
 that the system does not attempt to boot the SoC.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I165de0aa6c7185c8087328a427b0bb1eefb21e8d