gpio_commands: reduce number of vararg parameters
The extended GPIO state support command uses a lot of format elements
when trying to report properties of the GPIOs.

The upcoming source code processing scripts do not support more than
eight format elements per format string. On top of that there is no
need to even try to generate the string in case extended GPIO command
mode is supported.

Let's move generating the string with properties into the
conditionally compiled section, to make sure that it is
generated/processed only when needed, and let's use snprintf() first
to reduce the number of format elements in the following ccprintf
invocation.

BRANCH=cr50, cr50-mp
BUG=b:194964350
TEST=verified proper 'gpiog' command output when code is compiled both
     with and without CONFIG_CMD_GPIO_EXTENDED

Change-Id: I0836a350d1f787c84d2079f10de3652523a8a5a9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>