Trogdor: Create Trogdor baseboard
Move the common stuff to from board to baseboard to reduce the effort
of forking and maintaining derivatives.

This CL is intended to simply moving things around, not to address
other needs, like supporting a different EC.

Leave the following stuff in board:
 * Features for debugging
 * GPIO and interrupt handlers (likely will be changed)
 * BC 1.2 (already changed from Trogdor rev-0 and rev-1)
 * TCPC/MUX/PPC (will be changed)
 * Sensors (clamshell vs convertible)
 * ADC (detachable vs convertible)
 * LED/PWM
 * Battery

Others are moved to baseboard.

BRANCH=None
BUG=b:146237680
TEST=Build Trogdor and it boots. Should be good as there is no logic
change.

Change-Id: I2e4cd76d18f8739b5b7d5b1dae67c13e038b4480
Signed-off-by: Wai-Hong Tam <waihong@google.com>