ioexpander: add alternate group support
dalboz has two kind of daughter boards as options. Type A+C
board and Type A+HDMI board has different ioexpander chip
and the same signal names are used on HW schematics.

This patch expend a altgrp member for ioex_info and want
to add alternate support for this kind of configuration.

Change-Id: I0e05a8221ce3f136fe052a8370826017b3faf3f4
Signed-off-by: Paul Ma <magf@bitland.crop-partner.google.com>