stm32: Allow more accurate watchdog timing
For chips that specify an LSI clock frequency,
allow more accurate watchdog timing.

For the stm32h7, the watchdog was requested for 10 seconds,
but ended up being around 19 seconds. Meanwhile, the watchdog
helper messages was prompting around 9 seconds (using hsi).
This is because the watchdog code assumes a worst case clock
speed of 56KHz. Although this calculation is smart and works
for most cases/chips, I feel this can be confusing.
We need predicatability.

BRANCH=none
BUG=b:151528646
TEST=Tested on nucleo that reset occurred
     between waitms 9700 and waitms 10000.

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: I9ec59603d88811c0312549077336bc0541b81c00