Update patch set 5
Patch Set 5:

I will leave the decision of the alternate group for the others.  I did want to comment there is a common.h mechanism in place for the weak instances here that tends to document why it is being used.  Please use __overridable for the default definition, __override for the override definition and __override_proto to document the function can be overriden in the header prototype.

Patch-set: 5
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>