TCPMv1: Set initial data role based on power role
Set initial data role of the PD based on the current power role.
- Initial data role for sink is UFP
- Initial data role for source is DFP

Reference: USB Type-C cable and connector specification, Release 2.0
2.3.3 Initial Power (Source-to-Sink) Detection and Establishing the
      Data (Host-to-Device) Relationship

Once initial connection is established, the Source supplies VBUS and
behaves as a DFP, and the Sink consumes VBUS and behaves as a UFP.
USB PD, when supported by both ports, may then be used to independently
swap both the power and data roles of the ports.

BUG=b:151653348
BRANCH=none
TEST=Manually tested on Volteer
     Docks with power, do not change the data role on cold reset

Change-Id: I69e01f5b98a656394d61ee6cf657ea63a570abd8
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>