tcpmv2: cleanup auto discharge disconnect
This CL is based on the inability to boot trembyle when a
battery is not connected.  The failure looks like it is always
right after AutoDischargeDisconnect has been set.  I asked
Nuvoton if we could go back to using ForcedDischarge and
veer away from AutoDischargeDisconnect and I was told no
that it is now required.

My first step was to determine how close the existing
AutoDischargeDisconnect was to the spec and found enough
difference that I wanted to get that in line before
trying to pinpoint the issues causing this problem.

BUG=none
BRANCH=none
TEST=get tcpmv2 to work without a battery attached

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I44b86ae5dfcf6b547c742c1af0228a0ed8e3520d