Update patch set 2
Patch Set 2: Commit-Queue+2

Chris & I chatted over IM to see if we could get around this one-byte read restriction.

The idea was to use Popen.communicate with timeout set to a reasonably small value, polling in this manner over & over. This has a situation where we could run into deadlock if Python's call to read(2) was continually interrupted by SIGALRM and we never got to reading any bytes.

Another idea would be to move the tests to a process pool runner, and invoke a single program to run the tests instead of relying on the Makefile for parallelism. This is a good idea, but beyond the scope of this CL.

Patch-set: 2
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2