Update patch set 2
Patch Set 2:

Wanted to run without AutoDischargeDisconnect.  Amit from Nuvoton indicates:

"The NCT3807 is following the TCPC spec, therefore there is no way to run without AutoDischargeDisconnect."

This does not feel right with this being a requirement for a functional implementation.  Since the ForcedDischarge is still in the register set, it feels like this is required to be implemented by the TCPC but not required for use by the TCPM.

Patch-set: 2