Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> Wanted to run without AutoDischargeDisconnect.  Amit from Nuvoton indicates:
> 
> "The NCT3807 is following the TCPC spec, therefore there is no way to run without AutoDischargeDisconnect."
> 
> This does not feel right with this being a requirement for a functional implementation.  Since the ForcedDischarge is still in the register set, it feels like this is required to be implemented by the TCPC but not required for use by the TCPM.

It does seem weird that that TPCP would always require a bit to be set and not set it itself. Seems like a weird interface.

However, I do think we should set the auto discharge appropriately since it is called out in the Spec to do, and it is probably better from a compliance point of view too (we won't have floating Vbus on the connector)

Patch-set: 2