Update patch set 2
Patch Set 2:

Looks like this is exceeding the timeout on one test for some of the CQ builders. Specifically, when we have a lot of output, I'm suspecting we're hitting 64K in the pipe, and we have to wait for the test runner to catch up. Maybe I can be a little more clever.

Patch-set: 2
Reviewer: Gerrit User 1330845 <1330845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0