dedede: Use internal PU for MKBP_INT_L
Previous x86 designs have the EC_AP_MKBP_INT_L signal as an open drain
signal, but we don't have that configuration.  This commit adds the
ability for the EC to apply and remove an internal pullup when the SoC
turns on and off.

BUG=b:151680590
BRANCH=None
TEST=Build and flash waddledoo, verify that EC_MKBP_INT_L is asserted
when the AP shuts down and is de-asserted when the AP turns on.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I76a772f88ab8af1a9c7c699e5922eb86b926ae86