cmsg.c, console packet mode driver
This patch adds a console driver for packet mode.

The driver processes information prepared by util_precomplie.py when
processing printf invocations in the source code. Each invocation is
replaced by a cmsgX function, where X is the number of format
arguments.

cmsgX functions prepare an array of parameters of size X and invoke a
common function, passing it the array and an integer value, consisting
of up to 8 4 bit fields, describing the parameters.

Since both console drivers need to be able to filter logical console
channels, the channel_mask variable is made global.

BUG=b:149964350
TEST=with the rest of patches applied and packet mode enabled,
     verified proper operation of both Cr50 consoles (USB and UART).

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I1f6ef5ea50bffbe14d3e3850fff0191c54f37033