Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> Hi Amit, please could you take a look at tcpci_tcpc_set_connection() here and see how it aligns with what the NCT3807 expects?

As far as I understand  the first part (when DRP is set) looks o.k. 
But I’m not sure I understand why when working with no DRP you set bot CC to the same pull value, unless this code is called only when there is NO connection and DRP is not used.

Patch-set: 6