Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 5:
> > 
> > Hi Amit, please could you take a look at tcpci_tcpc_set_connection() here and see how it aligns with what the NCT3807 expects?
> 
> As far as I understand  the first part (when DRP is set) looks o.k. 
> But I’m not sure I understand why when working with no DRP you set bot CC to the same pull value, unless this code is called only when there is NO connection and DRP is not used.

I recently changed the code to set CC lines when the disconnect occurs and it is no longer needed here. Thank you for locating this for me.  The path is when we have turned off the application processor and no longer allow DRP.  When we disconnect a device it can get a false connection but DRP will not be enabled, so this path removes the alert so we don't get an alert storm.  I will remove the set_cc here.

Patch-set: 6