Update patch set 18
Patch Set 18:

(1 comment)

What I understand is how the device handles the WAIT message is in the scope of the Device Policy Manager. Our stack can have its own policy, like Sam's proposal of retrying 5 time.

The spec just states "This can occur in any case where the device receiving the PR_Swap Message needs to evaluate the request further e.g. by requesting Capabilities from the originator of the PR_Swap Message. Once it has completed this evaluation one of the Port Partners Should initiate the Power Role Swap process again by sending a PR_Swap Message."

The keyword is "one of the Port Partners". Sam's approach is our stack initiates the process again. And some other brand's approach may be just waiting the partner to initiate the process. In this case, our stack works robustly.

Aseda's previous comment about getting another swap request during the retry is a real concern. As either of the port partners can initiate the process again, it is expected that the partner may send the swap request. The current implementation of handling the retry in the Policy Engine seems not handling this case, as no way to back to PE_SRC_Ready/PE_SNK_Ready; in contrast, it will result a soft reset as an unexpected message is received.

A proper implementation may be doing the retry mechanism in the Device Policy Manager, instead of the Policy Engine. In the Policy Engine, it simply goes back to PE_SRC_Ready/PE_SNK_Ready state when it receives the Wait message and then signals the Device Policy Manager, which makes the decision to do another retry or timeout.

Patch-set: 18
CC: Gerrit User 1000407 <1000407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>