TCPMv2: Restore PD state after Sysjump
When a PD explicit contract is negotiated in RO,
the contract is maintained while performing a
sysjump to RW. This is done by serializing
the PD Power, Data, and VCONN Roles, along with
the explicit contract flag. After jumping to RW,
deserialization is performed by restoring the
power roles and setting the explicit contract
flag.

BUG=b:152350558,b:152027807,b:152967274
BRANCH=none
TEST=make -j buildall
	Manual tests:
	Used total phase to verify that charging voltage
	was maintained across sysjump from RO to RW and
	that DP was reestablished when a dock was used.

	Tested that Kohaku booted properly without a
	battery and a charger connected to port 0. This
	test was repeated with a charger connected to port 1.

Signed-off-by: Sam Hurst <shurst@google.com>

Change-Id: I349c41f2279e9af9830564d44ac73ad8435f1f80