Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> > Patch Set 14:
> > 
> > > Patch Set 14:
> > > 
> > > > Patch Set 14:
> > > > 
> > > > This tool is growing to the point we should consider migrating to Python before it gets too late.
> > > 
> > > Noted, but this tool has proven to be critical in validating no-functional-change CLs. Blocking these shell upgrades only slows productivity.
> > 
> > Yeah, but maybe open a P3 bug so we can start tracking migration. This tool only appeared a tiny bit ago, shouldn't be too hard.
> 
> Hmm, I'm prob being narrow minded. What would Python bring to the table here? Just better arg parsing?

One of the main things is testing. It's easy to write tests for python, but a lot harder for shell.

There's also the rationale in https://google.github.io/styleguide/shellguide.html#when-to-use-shell

When Craig and I discussed this in person I said that it's ok to land these for now since they're pretty much complete, but we should try to migrate to python in the future. My fault for writing it in shell to being with. 😊

Patch-set: 14
Reviewer: Gerrit User 1314845 <1314845@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>