Update patch set 9
Patch Set 9:

Ah, that somewhat defeats the purpose of the LED blinking, as a sign of life.
I use this to know when the system has booted and if the system is still ticking.

Is your goal to also ensure that the LEDs can still be used for other purposes?
If so, I don't need to hog all the LEDs. I could just make this blink 1 of the 3 LEDs.

Patch-set: 9
Reviewer: Gerrit User 1334530 <1334530@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0