TCPMv2: Set dual-role state basing on chipset state
Set DRP state during TCPMv2 initialization, basing on current chipset
state.

This was found when running master branch on nocturne. When EC jumps
from RO to RW and chipset state is S0 already, then it is expected that
DRP state is PD_DRP_TOGGLE_ON but actually the state is
PD_DRP_TOGGLE_OFF. DRP state is changed only during chipset power state
change, so it was necessary to shutdown and power on chromebook in
order to get USB-C devices working.

BUG=b:151301058
BRANCH=none
TEST=Run ChromeOS
     Plug some USB-C device and make sure device is working
     Reboot EC from EC console (EC should start from RO image)
     Run board
     Check if USB-C device is working. Do not replug it or reboot board

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: If83a463aabf8ee6493ba3dba44ad8829539c90cc